+++ Solutions/Trend Micro Vision One/Analytic Rules/Create Incident for XDR Alerts (Medium & Low).yaml
@@ -1,8 +1,9 @@
-﻿id: 0febd8cc-1b8d-45ed-87b3-e1e8a57d14cd
-name: Create Incident for XDR Alerts (Critical & High)
+﻿id: 00282588-11e7-436d-90e8-011256c3c691
+name: Create Incident for XDR Alerts (Medium & Low)
 description: |
-  'This Query creates an incident based on Trend Micro XDR Workbench Alerts and maps the impacted entities for Microsoft Sentinel usage. (Critical & High Serverity Alerts)'
+  'This Query creates an incident based on Trend Micro XDR Workbench Alerts and maps the impacted entities for Microsoft Sentinel usage. (Medium & Low Serverity Alerts)'
 severity: High
+status: Available
 requiredDataConnectors:
   - connectorId: TrendMicroXDR
     dataTypes:
@@ -17,7 +18,7 @@ query: |
 
   let threshold = 100;
   TrendMicro_XDR_CL        
-  | where modelSeverity_s == 'high' or modelSeverity_s  == 'critical'
+  | where modelSeverity_s == 'medium' or modelSeverity_s  == 'low'
   | extend AccountCustomEntity = impactScope_account_s, HostCustomEntity = impactScope_hostname_s, IPCustomEntity = impactScope_host_s
 entityMappings:
   - entityType: Account
