+++ Solutions/Trend Micro Vision One/Data Connectors/AzureFunctionTrendMicroXDR/shared_code/data/file_sub_true_type_mapping.json
