+++ Solutions/Trend Micro Vision One/Data Connectors/TrendMicroXDR.json
@@ -63,7 +63,7 @@
     ],
     "availability": {
         "status": 1,
-        "isPreview": true
+        "isPreview": false
     },
     "permissions": {
         "resourceProvider": [
