+++ Solutions/Trend Micro Vision One/Data Connectors/azuredeploy_TrendMicroVisionOne_API_FunctionApp.json
