+++ Solutions/Trend Micro Vision One/Data/Solution_TrendMicroVisionOneTemplateSpec.json
@@ -1,0 +1,22 @@
+{
+  "Name": "Trend Micro Vision One",
+  "Author": "Trend Micro",
+  "Logo": "<img src=\"https://raw.githubusercontent.com/Azure/Azure-Sentinel/master/Logos/Trend_Micro_Logo.svg\" width=\"75px\" height=\"75px\">",
+  "Description": "The [Trend Micro XDR](https://www.trendmicro.com/en_us/business/products/detection-response/xdr.html) solution for Microsoft Sentinel enables you to ingest security alerts and detection data reported in the Trend Micro XDR platform into Microsoft Sentinel. ",
+  "Data Connectors": [
+    "Data Connectors/TrendMicroXDR.json"
+  ],
+  "Workbooks": [
+  "Workbooks/TrendMicroXDROverview.json" 
+  ],
+"Analytic Rules": [
+"Analytic Rules/Create Incident for XDR Alerts (Critical & High).yaml",
+"Analytic Rules/Create Incident for XDR Alerts (Medium & Low).yaml"
+],
+
+  "BasePath": "C:\\GitHub\\Azure-Sentinel\\Solutions\\Trend Micro Vision One",
+  "Version": "2.0.0",
+  "Metadata": "SolutionMetadata.json",
+ "TemplateSpec": true,
+  "Is1PConnector": false
+}
