+++ Solutions/Trend Micro Vision One/SolutionMetadata.json
@@ -1,0 +1,19 @@
+{
+	"publisherId": "azuresentinel",
+	"offerId": "azure-sentinel-solution-isvtesting21",
+	"firstPublishDate": "2022-05-11",
+	"providers": ["Trend Micro"],
+	"categories": {
+		"domains" : ["Security - Threat Protection"],
+		"verticals": []
+	},
+	"support": {
+	  "name": "Trend Micro",
+	  "tier": "Partner",
+	  "link": "https://success.trendmicro.com/dcx/s/?language=en_US"
+	}
+}
+
+
+
+
